repo_stats/requirements.txt
python-sonarqube-api == 2.0.3